+++ src/hooks/state/index.ts
@@ -1,1 +1,2 @@
 export { useModalState } from './useModalState';
+export { useToggle } from './useToggle';